requirements.txt
django==1.11
djangorestframework==3.9.4
requests
pypdf2

